entry: Redo recording of undo
Add a start/end way of capturing the begin and end of an undoable
action. This is necessary because the entry code distinguishes between
changing the contents and updating the cursor position on a very high
level.
And we want to capture the cursor position update in the undo
implementation.

As a nice side effect, the 2 functions are nicely exportable (even for
public API) should we want to. And I guess we might want to for at least
entry completion and comboboxes.