undo: Change semantics of gtk_undo_command_merge()
Merges always succeed. The fallback implementation will create a (newly
added) GtkUndoCommandChain with the 2 commands to be merged.

The merge function may still return NULL. NULL should be returned if
merging commands would create a no-op.
This is implemented in the entry command undo. You can test it by
entering a character and then typing backspace.