undo: Add a way to record multiple commands into one
This is a so far just a simple object that collects commands, merges
them and then combines them into a single GtkUndoCommandChain at the
end. It's supposed to be used with a mechanism that goes
  begin_user_action();
  /* do stuff */
  end_user_action();
The begin_user_action() would create the recorder, then all new commands
would be put into the recorder and end_user_action() would then finish
the recorder and extract one resulting command.