build: Do not force inclusion of gtk.h
GCC will not do the right thing, and it will just break the build when
trying to include gtk.h first.

We'll have to live with the warning from the compiler about a missing
gtk_init() — though it would be better not to have to init GTK at all to
generate the introspection data.

This commit unbreaks the build in GNOME Continuous introduced by commit
6838861d26bfec9af53438022bd1f5b60bab3bf4.