gtkplacessidebar: protect for checking a null event
We are using the current gtk event position to check if
the user clicked in the eject button of a row.
However, we are calling this in row-activated signal, which
can happens when the gtk+ mainloop already removed the "current"
event, and therefore we were trying to access a null event position
which crashes the sidebar.

To fix it check if we have as current event, and if not,
we will not check for the position. But it's better than crashing,
and shouldn't happen too much.

https://bugzilla.gnome.org/show_bug.cgi?id=754995