wip: quartz: use colorspace from primary display
Using the default RGB colorspace seems to go through colorspace
conversion on tested macbook retina devices. Instead, default to the
colorspace of the main display so that it does not incur the conversion
cost on every frame.

This probably requires the equivalent patch to cairo, I have not tested
without it yet.