wip: pixelcache: ensure we use CGLayer on quartz
This is a special case for quartz where we want to ensure that a
CGLayer is used for the pixel cache. create_similar_surface() is not
enough since only a few CGLayer backed surfaces seem possible
before we can’t create them anymore.

This requires cairo-quartz patches for
cairo_quartz_surface_create_cg_layer()