gtkhidingbox: simplify visibility management
Extract a function, refactor, and manage the widget visibility
and child visibility at once, and use that for further checks.

This will be useful on a later patch when we add a property to choose
between hiding widget from the start or from the end, so we can allocate
or not children just checking the child visibility.